add 3 phase threshold & changerate

+++ stm32_src/app/frame_encode.h
@@ -14,7 +14,7 @@ uint16_t dev_info_encode(uint8_t *data, uint8_t errorcode, uint8_t * version, ui
 uint16_t frame_collection_cycle_data_encode(uint8_t *data, uint8_t errcode, uint8_t type, uint16_t cycle);
 
 /*周期数据编码*/
-uint16_t current_cycle_data_encode(uint8_t *data, uint8_t errorcode, uint8_t send_data_type,  uint8_t channel_count ,uint8_t channel_1, uint32_t ch1_current, uint8_t channel_2,uint32_t ch2_current, uint32_t timestamp);
+uint16_t current_cycle_data_encode(uint8_t *data, uint8_t errorcode, uint8_t send_data_type,  uint8_t channel_count ,uint8_t channel_1, float ch1_current, uint8_t channel_2,float ch2_current, uint32_t timestamp);
 
 uint16_t dip_angle_cycle_data_encode(uint8_t *data, uint8_t errorcode, uint32_t timestamp, float course_angle, float pitch_angle, float roll_angle);
 /*突变数据编码*/
